fix(wallet)_: logic for resolving swap between tokens updated
Based on what's said here https://developers.paraswap.network/api/get-tokens-list
```
ParaSwap's API is not limited to the tokens that are listed on the tokens endpoint. Tokens endpoint is merely a list of tokens curated by ParaSwap. This list is not actively maintained and will soon be deprecated. To use a more up-to-date list of tokens it is suggested to use token-lists.
```

We should not take tokens announced in that list as the only tokens available for swap, but try with any token the Status app supports and the Router will throw an error if swapping is not possible.